refactor: restructure plantio feature with new schema and form handling
- Removed the old usePlantioPayload hook and its related logic.
- Introduced new cadastro-plantio.schema.ts for form validation using Zod.
- Created formatPlantioForm function to format data for AI processing.
- Added new IA API schemas for plantio data handling.
- Implemented plantio actions for fetching habilidades and plantas.
- Developed a new CadastroPlantioForm component with improved field components.
- Added fields for ambiente, planta selection, and additional information.
- Integrated React Query for fetching habilidades dynamically.
- Updated styles for input fields to enhance user experience.
- Updated yarn.lock to use the latest versions of @tanstack/query-core and @tanstack/react-query.
