+++ package.json
@@ -14,7 +14,7 @@
         "@radix-ui/react-label": "^2.1.6",
         "@radix-ui/react-select": "^2.2.5",
         "@radix-ui/react-slot": "^1.2.2",
-        "@tanstack/react-query": "^5.76.1",
+        "@tanstack/react-query": "^5.80.7",
         "class-variance-authority": "^0.7.1",
         "clsx": "^2.1.1",
         "framer-motion": "^12.6.3",
