+++ src/features/cadastro-plantio/ui/SubmittedJsonDisplay.tsx
