+++ src/features/cadastro-plantio/ui/fields/styles.module.scss
@@ -1,0 +1,9 @@
+.numberInput{
+    &::-webkit-outer-spin-button,
+    &::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+    }
+    -moz-appearance: textfield;
+    appearance: textfield;
+}
